Added a QuizEntryDeserializer.java file to allow custom mapping of incoming json objects to native java QuizEntry objects, including tag mapping.

+++ app/src/main/java/com/example/learnquest/AppState/App.java
@@ -12,8 +12,8 @@ import com.example.learnquest.Utils.database.SupabaseClient;
  */
 public class App {
 
-    public static Short userID;
-    public static Short groupID;
+    public static Integer userID;
+    public static Integer groupID;
     public static Context applicationContext;
 
     public static void setApplicationContext(Context context){
